- replacing original sleep staging example by new one on sequences - fixing formatting in sleep staging example - adding SequenceSampler to docs
